refactor: seperate rendering module
