+++ crates/uzumaki/src/element.rs
@@ -1,23 +1,18 @@
 use std::cell::Cell;
-use std::collections::HashMap;
 use std::rc::Rc;
 
-use cosmic_text::Attrs;
 use slab::Slab;
 use unicode_segmentation::UnicodeSegmentation;
-use vello::Scene;
-use vello::kurbo::{Affine, Rect, RoundedRect, RoundedRectRadii};
-use vello::peniko::{Color as VelloColor, Fill};
 
 use crate::cursor::CursorIcon;
-use crate::element::input::{InputRenderInfo, compute_selection_rects};
 use crate::input::{BaseInputState, RangeProvider};
 use crate::interactivity::{HitTestState, HitboxStore, Interactivity};
 use crate::selection::{DomSelection, SelectionRange};
-use crate::style::{Bounds, Color, Style};
+use crate::style::{Bounds, Style};
 use crate::text::TextRenderer;
 
 pub mod input;
+pub mod render;
 pub mod text;
 pub mod view;
 
@@ -713,7 +708,7 @@ impl ElementTree {
                         height: taffy::AvailableSpace::Definite(height),
                     },
                     |known_dimensions, available_space, _node_id, node_context, _style| {
-                        Self::measure(
+                        render::measure(
                             text_renderer,
                             known_dimensions,
                             available_space,
@@ -749,554 +744,6 @@ impl ElementTree {
     /// Set the active node (mouse down on an element).
     pub fn set_active(&mut self, node_id: Option<NodeId>) {
         self.hit_state.active_node = node_id;
-    }
-
-    /// Render the DOM tree into the scene. Also rebuilds hitboxes and scroll thumbs.
-    pub fn render(&mut self, scene: &mut Scene, text_renderer: &mut TextRenderer, scale: f64) {
-        self.hitbox_store.clear();
-        self.scroll_thumbs.clear();
-        self.build_text_select_runs();
-
-        if let Some(root) = self.root {
-            self.render_tree(scene, text_renderer, root, scale);
-        }
-    }
-
-    fn render_tree(
-        &mut self,
-        scene: &mut Scene,
-        text_renderer: &mut TextRenderer,
-        root_id: NodeId,
-        scale: f64,
-    ) {
-        // Collect render info for all nodes in DFS order
-        struct RenderInfo {
-            node_id: NodeId,
-            x: f64,
-            y: f64,
-            w: f64,
-            h: f64,
-            style: Box<Style>,
-            text: Option<(String, f32, Color)>,
-            needs_hitbox: bool,
-            input: Option<InputRenderInfo>,
-        }
-
-        struct ThumbInfo {
-            node_id: NodeId,
-            view_x: f64,
-            view_y: f64,
-            view_w: f64,
-            view_h: f64,
-            scroll_offset_y: f32,
-            content_height: f32,
-            visible_height: f32,
-            thumb_hovered: bool,
-            mouse_in_view: bool,
-            scale: f64,
-        }
-
-        enum RenderCommand {
-            PaintNode(RenderInfo),
-            PushClip(Rect, f64),
-            PopClip,
-            PaintThumb(ThumbInfo),
-        }
-
-        enum StackItem {
-            Visit(NodeId, f64, f64, InheritedProperties),
-            PushClip(Rect, f64),
-            PopClip,
-            PaintThumb(ThumbInfo),
-        }
-
-        // Pre-compute per-node selection ranges for text selection painting
-        let text_sel_map = self.compute_text_selections_map();
-
-        let mut render_list: Vec<RenderCommand> = Vec::new();
-        let mut stack: Vec<StackItem> = vec![StackItem::Visit(
-            root_id,
-            0.0,
-            0.0,
-            InheritedProperties::default(),
-        )];
-
-        while let Some(item) = stack.pop() {
-            match item {
-                StackItem::PushClip(rect, s) => {
-                    render_list.push(RenderCommand::PushClip(rect, s));
-                    continue;
-                }
-                StackItem::PopClip => {
-                    render_list.push(RenderCommand::PopClip);
-                    continue;
-                }
-                StackItem::PaintThumb(info) => {
-                    render_list.push(RenderCommand::PaintThumb(info));
-                    continue;
-                }
-                StackItem::Visit(node_id, parent_x, parent_y, parent_inherited) => {
-                    // Extract all needed data from the node (immutable borrow scope)
-                    let (
-                        taffy_node,
-                        computed_style,
-                        text,
-                        input,
-                        needs_hitbox,
-                        is_scrollable,
-                        first_child,
-                        inherited,
-                    ) = {
-                        let node = &self.nodes[node_id];
-
-                        // Resolve inherited properties
-                        let mut inherited = parent_inherited.clone();
-                        if let Some(ts) = node.selectable {
-                            inherited.selectable = ts;
-                        }
-
-                        let taffy_node = node.taffy_node;
-                        let computed_style =
-                            node.interactivity
-                                .compute_style(&node.style, node_id, &self.hit_state);
-
-                        let text = node.behavior.as_text().map(|tc| {
-                            (
-                                tc.content.clone(),
-                                computed_style.text.font_size,
-                                computed_style.text.color,
-                            )
-                        });
-
-                        let input = node.behavior.as_input().map(|is| {
-                            let range = is.range();
-                            InputRenderInfo {
-                                display_text: is.display_text(),
-                                placeholder: is.placeholder.clone(),
-                                font_size: computed_style.text.font_size,
-                                text_color: computed_style.text.color,
-                                focused: is.focused,
-                                sel_start: range.start(),
-                                sel_end: range.end(),
-                                cursor_pos: range.active,
-                                scroll_offset: is.scroll_offset,
-                                scroll_offset_y: is.scroll_offset_y,
-                                blink_visible: is.blink_visible(self.window_focused),
-                                multiline: is.multiline,
-                            }
-                        });
-
-                        // Text nodes inside textSelect views need hitboxes for click-to-select
-                        let selectable_text = inherited.selectable && node.behavior.is_text();
-                        let needs_hitbox = node.interactivity.needs_hitbox() || selectable_text;
-                        let is_scrollable = node.scroll_state.is_some();
-                        let first_child = node.first_child;
-
-                        (
-                            taffy_node,
-                            computed_style,
-                            text,
-                            input,
-                            needs_hitbox,
-                            is_scrollable,
-                            first_child,
-                            inherited,
-                        )
-                    };
-                    // immutable borrow of self.nodes is now dropped
-
-                    let Ok(layout) = self.taffy.layout(taffy_node) else {
-                        continue;
-                    };
-
-                    let x = parent_x + layout.location.x as f64;
-                    let y = parent_y + layout.location.y as f64;
-                    let w = layout.size.width as f64;
-                    let h = layout.size.height as f64;
-
-                    // Compute scroll info and clamp offset (mutable borrow is safe now)
-                    let scroll_info = if is_scrollable {
-                        let content_height = layout.content_size.height;
-                        let visible_height = layout.size.height;
-                        let max_scroll = (content_height - visible_height).max(0.0);
-                        if let Some(ss) = self.nodes[node_id].scroll_state.as_mut()
-                            && ss.scroll_offset_y > max_scroll
-                        {
-                            ss.scroll_offset_y = max_scroll;
-                        }
-                        let clamped_offset = self.nodes[node_id]
-                            .scroll_state
-                            .as_ref()
-                            .map(|ss| ss.scroll_offset_y)
-                            .unwrap_or(0.0);
-                        Some((content_height, visible_height, clamped_offset))
-                    } else {
-                        None
-                    };
-
-                    // Collect children in order
-                    let mut children = Vec::new();
-                    let mut child = first_child;
-                    while let Some(child_id) = child {
-                        children.push(child_id);
-                        child = self.nodes[child_id].next_sibling;
-                    }
-
-                    if let Some((content_height, visible_height, clamped_offset)) = scroll_info {
-                        let overflows = content_height > visible_height;
-                        let thumb_hovered = self
-                            .scroll_drag
-                            .as_ref()
-                            .is_some_and(|d| d.node_id == node_id)
-                            || self.scroll_thumbs.iter().any(|t| {
-                                t.node_id == node_id
-                                    && self
-                                        .hit_state
-                                        .mouse_position
-                                        .is_some_and(|(mx, my)| t.thumb_bounds.contains(mx, my))
-                            });
-
-                        let mouse_in_view = self
-                            .scroll_drag
-                            .as_ref()
-                            .is_some_and(|d| d.node_id == node_id)
-                            || self.hit_state.mouse_position.is_some_and(|(mx, my)| {
-                                mx >= x && mx <= x + w && my >= y && my <= y + h
-                            });
-
-                        // Push in reverse order for LIFO stack:
-                        // 6. PaintThumb (last to execute)
-                        if overflows {
-                            stack.push(StackItem::PaintThumb(ThumbInfo {
-                                node_id,
-                                view_x: x,
-                                view_y: y,
-                                view_w: w,
-                                view_h: h,
-                                scroll_offset_y: clamped_offset,
-                                content_height,
-                                visible_height,
-                                thumb_hovered,
-                                mouse_in_view,
-                                scale,
-                            }));
-                        }
-                        // 5. PopClip
-                        stack.push(StackItem::PopClip);
-                        // 4-3. Children (reversed for correct order)
-                        for &child_id in children.iter().rev() {
-                            stack.push(StackItem::Visit(
-                                child_id,
-                                x,
-                                y - clamped_offset as f64,
-                                inherited.clone(),
-                            ));
-                        }
-                        // 2. PushClip
-                        let clip_rect = Rect::new(x, y, x + w, y + h);
-                        stack.push(StackItem::PushClip(clip_rect, scale));
-                    } else {
-                        // Normal (non-scrollable) node: push children
-                        for &child_id in children.iter().rev() {
-                            stack.push(StackItem::Visit(child_id, x, y, inherited.clone()));
-                        }
-                    }
-
-                    // 1. PaintNode (always first — the node's own bg/borders)
-                    render_list.push(RenderCommand::PaintNode(RenderInfo {
-                        node_id,
-                        x,
-                        y,
-                        w,
-                        h,
-                        style: Box::new(computed_style),
-                        text,
-                        needs_hitbox,
-                        input,
-                    }));
-                }
-            }
-        }
-
-        // Paint all commands in order
-        for cmd in &render_list {
-            match cmd {
-                RenderCommand::PaintNode(info) => {
-                    let bounds = Bounds::new(info.x, info.y, info.w, info.h);
-
-                    // Register hitbox if needed
-                    if info.needs_hitbox {
-                        let hitbox_id = self.hitbox_store.insert(info.node_id, bounds);
-                        self.nodes[info.node_id].interactivity.hitbox_id = Some(hitbox_id);
-                    }
-
-                    if let Some(input_info) = &info.input {
-                        let content_info = crate::element::input::paint_input(
-                            scene,
-                            text_renderer,
-                            bounds,
-                            &info.style,
-                            input_info,
-                            scale,
-                        );
-
-                        // Paint scrollbar for multiline inputs with overflow
-                        if let Some(ci) = content_info
-                            && ci.content_height > ci.visible_height
-                        {
-                            let mouse_in = self
-                                .scroll_drag
-                                .as_ref()
-                                .is_some_and(|d| d.node_id == info.node_id)
-                                || self
-                                    .hit_state
-                                    .mouse_position
-                                    .is_some_and(|(mx, my)| bounds.contains(mx, my));
-
-                            let thumb_width = 4.0;
-                            let thumb_margin = 4.0;
-                            let ratio = ci.visible_height / ci.content_height;
-                            let thumb_height = (bounds.height * ratio).max(24.0);
-                            let max_scroll = (ci.content_height - ci.visible_height).max(0.0);
-                            let scroll_ratio = if max_scroll > 0.0 {
-                                ci.scroll_offset_y / max_scroll
-                            } else {
-                                0.0
-                            };
-                            let thumb_y = bounds.y + scroll_ratio * (bounds.height - thumb_height);
-                            let thumb_x = bounds.x + bounds.width - thumb_width - thumb_margin;
-
-                            let thumb_bounds =
-                                Bounds::new(thumb_x, thumb_y, thumb_width, thumb_height);
-
-                            // Register for hit testing (drag + wheel)
-                            self.scroll_thumbs.push(ScrollThumbRect {
-                                node_id: info.node_id,
-                                thumb_bounds,
-                                view_bounds: bounds,
-                                content_height: ci.content_height as f32,
-                                visible_height: ci.visible_height as f32,
-                            });
-
-                            if mouse_in {
-                                let thumb_hovered = self
-                                    .scroll_drag
-                                    .as_ref()
-                                    .is_some_and(|d| d.node_id == info.node_id)
-                                    || self
-                                        .hit_state
-                                        .mouse_position
-                                        .is_some_and(|(mx, my)| thumb_bounds.contains(mx, my));
-                                let alpha = if thumb_hovered { 140u8 } else { 90u8 };
-                                let color = VelloColor::from_rgba8(255, 255, 255, alpha);
-                                let radius = thumb_width / 2.0;
-                                let rect = Rect::new(
-                                    thumb_x,
-                                    thumb_y,
-                                    thumb_x + thumb_width,
-                                    thumb_y + thumb_height,
-                                );
-                                let rounded = RoundedRect::from_rect(
-                                    rect,
-                                    RoundedRectRadii::from_single_radius(radius),
-                                );
-                                // Clip to input bounds
-                                let clip = Rect::new(
-                                    bounds.x,
-                                    bounds.y,
-                                    bounds.x + bounds.width,
-                                    bounds.y + bounds.height,
-                                );
-                                scene.push_clip_layer(Fill::NonZero, Affine::scale(scale), &clip);
-                                scene.fill(
-                                    Fill::NonZero,
-                                    Affine::scale(scale),
-                                    color,
-                                    None,
-                                    &rounded,
-                                );
-                                scene.pop_layer();
-                            }
-                        }
-                    } else if let Some((content, font_size, color)) = &info.text {
-                        let sel_range = text_sel_map.get(&info.node_id).copied();
-                        if sel_range.is_some() {
-                            // Text node with active selection: paint selection
-                            // highlight between background and text glyphs.
-                            info.style.paint(bounds, scene, scale, |scene| {
-                                if let Some((sel_start, sel_end)) = sel_range {
-                                    let positions = text_renderer.grapheme_positions_2d(
-                                        content,
-                                        *font_size,
-                                        Some(bounds.width as f32),
-                                    );
-                                    let line_height = (*font_size * 1.2).round();
-                                    let rects = compute_selection_rects(
-                                        &positions,
-                                        sel_start,
-                                        sel_end,
-                                        bounds.width,
-                                        line_height as f64,
-                                    );
-                                    let sel_color = VelloColor::from_rgba8(56, 121, 185, 128);
-                                    for [x1, y1, x2, y2] in rects {
-                                        scene.fill(
-                                            Fill::NonZero,
-                                            Affine::scale(scale),
-                                            sel_color,
-                                            None,
-                                            &Rect::new(
-                                                bounds.x + x1,
-                                                bounds.y + y1,
-                                                bounds.x + x2,
-                                                bounds.y + y2,
-                                            ),
-                                        );
-                                    }
-                                }
-                                text_renderer.draw_text(
-                                    scene,
-                                    content,
-                                    Attrs::new(),
-                                    *font_size,
-                                    bounds.width as f32,
-                                    bounds.height as f32,
-                                    (bounds.x as f32, bounds.y as f32),
-                                    color.to_vello(),
-                                    scale,
-                                );
-                            });
-                        } else {
-                            crate::element::text::paint_text(
-                                scene,
-                                text_renderer,
-                                bounds,
-                                &info.style,
-                                content,
-                                *font_size,
-                                *color,
-                                scale,
-                            );
-                        }
-                    } else {
-                        crate::element::view::paint_view(scene, bounds, &info.style, scale, |_| {});
-                    }
-                }
-                RenderCommand::PushClip(rect, s) => {
-                    scene.push_clip_layer(Fill::NonZero, Affine::scale(*s), rect);
-                }
-                RenderCommand::PopClip => {
-                    scene.pop_layer();
-                }
-                RenderCommand::PaintThumb(thumb) => {
-                    // Only show scrollbar when mouse is inside the scrollable node
-                    if !thumb.mouse_in_view {
-                        continue;
-                    }
-
-                    // Scrollbar thumb: 4px wide, 4px margin from right edge
-                    let thumb_width = 4.0;
-                    let thumb_margin = 4.0;
-
-                    let ratio = thumb.visible_height as f64 / thumb.content_height as f64;
-                    let thumb_height = (thumb.view_h * ratio).max(24.0);
-                    let track_height = thumb.view_h;
-                    let max_scroll = (thumb.content_height - thumb.visible_height).max(0.0) as f64;
-                    let scroll_ratio = if max_scroll > 0.0 {
-                        thumb.scroll_offset_y as f64 / max_scroll
-                    } else {
-                        0.0
-                    };
-                    let thumb_y = thumb.view_y + scroll_ratio * (track_height - thumb_height);
-                    let thumb_x = thumb.view_x + thumb.view_w - thumb_width - thumb_margin;
-
-                    let thumb_bounds = Bounds::new(thumb_x, thumb_y, thumb_width, thumb_height);
-
-                    // Register for hit testing
-                    self.scroll_thumbs.push(ScrollThumbRect {
-                        node_id: thumb.node_id,
-                        thumb_bounds,
-                        view_bounds: Bounds::new(
-                            thumb.view_x,
-                            thumb.view_y,
-                            thumb.view_w,
-                            thumb.view_h,
-                        ),
-                        content_height: thumb.content_height,
-                        visible_height: thumb.visible_height,
-                    });
-
-                    // Paint the thumb as a filled rounded rect
-                    let alpha = if thumb.thumb_hovered { 140u8 } else { 90u8 };
-                    let color = VelloColor::from_rgba8(255, 255, 255, alpha);
-                    let radius = thumb_width / 2.0;
-                    let rect = Rect::new(
-                        thumb_x,
-                        thumb_y,
-                        thumb_x + thumb_width,
-                        thumb_y + thumb_height,
-                    );
-                    let rounded =
-                        RoundedRect::from_rect(rect, RoundedRectRadii::from_single_radius(radius));
-                    scene.fill(
-                        Fill::NonZero,
-                        Affine::scale(thumb.scale),
-                        color,
-                        None,
-                        &rounded,
-                    );
-                }
-            }
-        }
-    }
-
-    fn measure(
-        text_renderer: &mut TextRenderer,
-        known_dimensions: taffy::Size<Option<f32>>,
-        available_space: taffy::Size<taffy::AvailableSpace>,
-        node_context: Option<&mut NodeContext>,
-    ) -> taffy::Size<f32> {
-        let default_size = taffy::Size {
-            width: known_dimensions.width.unwrap_or(0.0),
-            height: known_dimensions.height.unwrap_or(0.0),
-        };
-
-        let Some(ctx) = node_context else {
-            return default_size;
-        };
-
-        if ctx.is_input {
-            return taffy::Size {
-                width: known_dimensions
-                    .width
-                    .or_else(|| available_as_option(available_space.width))
-                    .unwrap_or(200.0),
-                height: known_dimensions
-                    .height
-                    .unwrap_or(ctx.font_size * 1.2 + 16.0),
-            };
-        }
-
-        if let Some(text) = &ctx.text {
-            let (measured_width, measured_height) = text_renderer.measure_text(
-                &text.content,
-                Attrs::new(),
-                ctx.font_size,
-                known_dimensions
-                    .width
-                    .or_else(|| available_as_option(available_space.width)),
-                known_dimensions
-                    .height
-                    .or_else(|| available_as_option(available_space.height)),
-            );
-
-            return taffy::Size {
-                width: measured_width,
-                height: measured_height,
-            };
-        }
-
-        default_size
     }
 
     /// Dispatch mouse down event to listeners on hovered elements.
@@ -1407,39 +854,6 @@ impl ElementTree {
                 stack.push((cid, resolved_ts, current_run));
             }
         }
-    }
-
-    /// Pre-compute per-text-node selection ranges for the current frame.
-    /// Returns a map from NodeId → (local_sel_start, local_sel_end) in grapheme units.
-    fn compute_text_selections_map(&self) -> HashMap<NodeId, (usize, usize)> {
-        let mut map = HashMap::new();
-        let Some(sel) = self.selection.get() else {
-            return map;
-        };
-        if sel.is_collapsed() {
-            return map;
-        }
-        let Some(run) = self
-            .selectable_text_runs
-            .iter()
-            .find(|r| r.root_id == sel.root)
-        else {
-            return map;
-        };
-        let sel_start = sel.start();
-        let sel_end = sel.end();
-        for entry in &run.entries {
-            let entry_end = entry.flat_start + entry.grapheme_count;
-            let local_start = sel_start.max(entry.flat_start);
-            let local_end = sel_end.min(entry_end);
-            if local_start < local_end {
-                map.insert(
-                    entry.node_id,
-                    (local_start - entry.flat_start, local_end - entry.flat_start),
-                );
-            }
-        }
-        map
     }
 
     /// Find the text run that contains a given text node.
@@ -1579,13 +993,6 @@ impl ElementTree {
     /// Clear the selection.
     pub fn clear_selection(&mut self) {
         self.selection.clear();
-    }
-}
-
-fn available_as_option(space: taffy::AvailableSpace) -> Option<f32> {
-    match space {
-        taffy::AvailableSpace::Definite(v) => Some(v),
-        _ => None,
     }
 }
